Max num Strategy only injects if body.length equals this.times.
This is usefull because on the ocasion of a race conditions append may be called two times before setStatus is called. The strategy will then aggregate once with this fix instead of twice. It will contain 1 extra event but it will not trigger the aggregation twice.

The second aggregation could trigger a false aggregation if between the first and the second one an append is called.